🩹 fix / 파일 구조 공통화
