+++ src/components/noteComponents/TopicCategoryForm.tsx
@@ -2,7 +2,7 @@
 
 import { TOPIC_CATEGORIES } from '@/constants/openai/category';
 import { useNoteStore } from '@/store/note/noteStore';
-
+import Text from '../common/Text';
 interface Props {
   onSelectCategory: (topic: string) => void;
 }
@@ -10,28 +10,33 @@ interface Props {
 const TopicCategoryForm = ({ onSelectCategory }: Props) => {
   const { selectedTopic } = useNoteStore();
   return (
-    <div className="p-5">
-      <p className="text-[22px] font-semibold p-5">
-        걱정은 어디에서 왔나요?
-        <br />
-        어떤 주제인가요?
-      </p>
-
-      <div className="flex flex-wrap gap-2 p-5">
-        {TOPIC_CATEGORIES.map((topic) => {
-          const isTopic = selectedTopic === topic;
-          return (
-            <button
-              key={topic}
-              onClick={() => onSelectCategory(topic)}
-              className={`px-3 py-[6px] rounded-full text-label-normal text-[16px] gap-1 border-[1px] border-line-normal  flex flex-col items-center justify-center
+    <div>
+      <div className="w-full h-auto px-5 py-2">
+        <Text variant="heading3" color="label-normal">
+          걱정은 어디에서 왔나요?
+          <br />
+          어떤 주제인가요?
+        </Text>
+      </div>
+      <div className="p-5">
+        <div className="flex flex-wrap gap-2">
+          {TOPIC_CATEGORIES.map((topic) => {
+            const isTopic = selectedTopic === topic;
+            return (
+              <button
+                key={topic}
+                onClick={() => onSelectCategory(topic)}
+                className={`px-3 py-[6px] rounded-[16px]  border-[1px] border-line-normal  flex flex-col items-center justify-center
                 ${!isTopic ? 'bg-white text-black' : 'bg-primary-3 text-white'}
               `}
-            >
-              {topic}
-            </button>
-          );
-        })}
+              >
+                <Text variant="body2" color="label-neutral">
+                  {topic}
+                </Text>
+              </button>
+            );
+          })}
+        </div>
       </div>
     </div>
   );
