+++ src/hooks/ranking/useAnaylsisTrendWeek.ts
@@ -1,5 +1,5 @@
 import { analyzeWeeklyCategoryTrends } from '@/app/utils/ranking/DataFetch';
-import { useUserStore } from '@/store/store';
+import { useUserStore } from '@/store/auth/store';
 import { AnalysisWeekTrendsResult } from '@/types/ranking/types';
 import { useEffect, useState } from 'react';
